Detect when an instruction pops from an empty main stack.
A buffer zone is created before the ^! stack. If, at the start of the
main loop, the stack pointer (rsp) is greater than the base pointer
(rbp), it is detected and reported as an error. Note that instructions
like `!` and `@` will not be detected, but due to the buffer space,
have no adverse effects since the program owns the stack memory used.